// apps/frontend/src/gen/vibesync/common/v1/common_pb.ts
export enum SystemRole {
  /**
   * @generated from enum value: SYSTEM_ROLE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: SYSTEM_ROLE_GUEST = 1;
   */
  GUEST = 1,

  /**
   * @generated from enum value: SYSTEM_ROLE_USER = 2;
   */
  USER = 2,

  /**
   * @generated from enum value: SYSTEM_ROLE_MODERATOR = 3;
   */
  MODERATOR = 3,

  /**
   * @generated from enum value: SYSTEM_ROLE_ADMINISTRATOR = 4;
   */
  ADMINISTRATOR = 4,
}