// apps/frontend/src/gen/vibesync/auth/v1/auth_pb.ts
export class LinkedProvider extends Message<LinkedProvider> {
  /**
   * Provider name ("spotify" or "google").
   *
   * @generated from field: string provider = 1;
   */
  provider = "";

  /**
   * @generated from field: google.protobuf.Timestamp linked_at = 2;
   */
  linkedAt?: Timestamp;

  constructor(data?: PartialMessage<LinkedProvider>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "vibesync.auth.v1.LinkedProvider";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "provider", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "linked_at", kind: "message", T: Timestamp },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): LinkedProvider {
    return new LinkedProvider().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): LinkedProvider {
    return new LinkedProvider().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): LinkedProvider {
    return new LinkedProvider().fromJsonString(jsonString, options);
  }

  static equals(a: LinkedProvider | PlainMessage<LinkedProvider> | undefined, b: LinkedProvider | PlainMessage<LinkedProvider> | undefined): boolean {
    return proto3.util.equals(LinkedProvider, a, b);
  }
}